Lots of initial work on GNUstep support

+++ Spec/Support/ObjectWithProperty.h
@@ -1,10 +1,9 @@
 #import <Foundation/Foundation.h>
-#import <CoreGraphics/CoreGraphics.h>
 #import "CedarObservedObject.h"
 
 @interface ObjectWithProperty : NSObject <CedarObservedObject>
 
-@property (nonatomic, assign) CGFloat floatProperty;
-@property (nonatomic, assign) CGFloat manualFloatProperty;
+@property (nonatomic, assign) float floatProperty;
+@property (nonatomic, assign) float manualFloatProperty;
 
 @end
